backend/requirements.txt
django===4.2.6
djangorestframework===3.14.0
django-cors-headers===4.3.1
python-decouple===3.8
# django-recaptcha===4.0.0
# captcha===0.5.0
# django-ranged-response===0.2.0 
# django-simple-captcha===0.5.20
